# core-common-modelhub-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-modelhub
  description: ModelHub Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/modelhub
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/fields/{id}:
    get:
      tags:
        - Virtual Entity Fields
      summary: Get a field by ID
      description: Returns a virtual entity field definition by its ID
      operationId: getFieldById
      parameters:
        - name: id
          in: path
          description: ID of the field to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualEntityFieldDto'
        '404':
          description: Field not found
    put:
      tags:
        - Virtual Entity Fields
      summary: Update a field
      description: Updates an existing field definition
      operationId: updateField
      parameters:
        - name: id
          in: path
          description: ID of the field to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VirtualEntityFieldDto'
        required: true
      responses:
        '200':
          description: Field updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualEntityFieldDto'
        '400':
          description: Invalid input
        '404':
          description: Field not found
    delete:
      tags:
        - Virtual Entity Fields
      summary: Delete a field
      description: Deletes a field definition
      operationId: deleteField
      parameters:
        - name: id
          in: path
          description: ID of the field to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Field deleted successfully
        '404':
          description: Field not found
...
  /api/v1/fields:
    post:
      tags:
        - Virtual Entity Fields
      summary: Create a new field
      description: Creates a new field definition for a virtual entity
      operationId: createField
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VirtualEntityFieldDto'
        required: true
      responses:
        '201':
          description: Field created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualEntityFieldDto'
        '400':
          description: Invalid input
        '404':
          description: Entity not found
...
  /api/v1/fields/entity/{entityId}:
    get:
      tags:
        - Virtual Entity Fields
      summary: Get all fields for an entity
      description: Returns a list of all fields defined for a specific virtual entity
      operationId: getFieldsByEntityId
      parameters:
        - name: entityId
          in: path
          description: ID of the entity to retrieve fields for
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VirtualEntityFieldDto'
        '404':
          description: Entity not found
...
components:
  schemas:
    VirtualEntityFieldDto:
      required:
        - entityId
        - fieldKey
        - fieldType
      type: object
      properties:
        id:
          type: string
          description: Unique identifier
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        entityId:
          type: string
          description: Parent entity ID
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        fieldKey:
          type: string
          description: Field key (used in code/API)
          example: firstName
        fieldLabel:
          type: string
          description: Field label (displayed to users)
          example: First Name
        fieldType:
          type: string
          description: Field data type
          example: string
        required:
          type: boolean
          description: Whether the field is required
          example: true
        options:
          type: object
          additionalProperties:
            type: object
            description: Additional field options (validation, UI hints, etc.)
          description: Additional field options (validation, UI hints, etc.)
        orderIndex:
          type: integer
          description: Display order index
          format: int32
          example: 1
        createdBy:
          type: string
          description: Created by user
          example: admin
        createdAt:
          type: string
          description: Creation timestamp
          format: date-time
        updatedBy:
          type: string
          description: Last updated by user
          example: admin
        updatedAt:
          type: string
          description: Last update timestamp
          format: date-time
      description: Field to create
...